requirements.txt
Django==4.2.6
djangorestframework==3.14.0
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
scipy==1.11.4
statsmodels==0.14.1
